Support write composite and array types in text format.

+++ lib/inc/pfs/IResult.h
@@ -9,10 +9,10 @@ class IResult
 {
 public:
     virtual ~IResult() = default;
-    virtual int rows() const noexcept = 0;
-    virtual int columns() const noexcept = 0;
-    virtual bool isNull(int row, int col) const noexcept = 0;
-    virtual const char * getValue(int row, int col) const noexcept = 0;
-    virtual int getLength(int row, int col) const noexcept = 0;
+    virtual size_t rows() const noexcept = 0;
+    virtual size_t columns() const noexcept = 0;
+    virtual bool isNull(size_t row, size_t col) const noexcept = 0;
+    virtual const char * getValue(size_t row, size_t col) const noexcept = 0;
+    virtual size_t getLength(size_t row, size_t col) const noexcept = 0;
 };
 } // namespace pfs
